Add abstract `ActionableElement`, an `Element` sub-type with action ID, add the `Image` elements Refs #7
## Description
This set of changes add the abstract `Image` `ActionableElement` and its concrete `UrlImage` and `SlackFileImage` elements. It also introduces the abstract `ActionableElement` element, a sub-type of `Element` with an action ID. Most elements inherit from it, except for certain elements like the new `Image` elements being introduced with these changes, which inherit directly from `Element` are `Image` elements are not actionable.

## Changes in this Pull Request
- add `ActionableElement`, an abstract `Element` with an action ID
- remove `actionID` from `Element`: not all elements are actionable, so `actionID` is now only defined in the new `ActionableElement` type, a sub-type of `Element`
- update existing sub-types of `Element` to inherit from `ActionableElement`, as all existing types that use `Element` are `ActionableElement`s
- add the `Image` element a non-actionable `Element`